+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "ast": "rollup -c && node src/test.js",
     "build": "rollup -c",
-    "dev": "rollup -c && node ./test"
+    "dev": "rollup -c && node ./test",
+    "test": "mocha"
   },
   "repository": {
     "type": "git",
@@ -43,6 +44,8 @@
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-es2015-rollup": "^3.0.0",
+    "chai": "^4.1.2",
+    "mocha": "^5.1.1",
     "noshjs": "^0.3.0",
     "rollup": "^0.57.1",
     "rollup-plugin-babel": "^3.0.3"
