+++ test/scss/selector.scss
@@ -1,0 +1,7 @@
+#app {
+  color: red;
+
+  div {
+    width: 100px;
+  }
+}
