+++ test/stylus/selector.styl
@@ -1,0 +1,5 @@
+#app
+  color: red
+
+  div
+    width: 100px
